func getData devided in 2: request and parsing (ParseHandler added)
