+++ UniversityHW/DataHandlers/DataHandler.swift
@@ -15,17 +15,17 @@ class  DataHandler {
     private init() {
     }
 
-    let apiHandler = APIHandler()
-    //let parseHandler = ParseHandler()
+    //let apiHandler = APIHandler()
+    let parseHandler = ParseHandler()
     //let storageHandler = StorageHandler(storage: StorageType)
 
     let urlString = "https://jsonplaceholder.typicode.com/users"
 
-    func handle(completion: @escaping ([Student]) -> Void) {
-        apiHandler.getData(urlString: urlString, completion: completion)
+//    func handle(completion: @escaping ([Student]) -> Void) {
+//        apiHandler.getData(urlString: urlString, completion: completion)
 
         //let dataRecieved = apiHandler.requestDataToAPI()
         //let arrayOfStudents = parseHandler.parse(data: dataRecieved)
         //storageHandler.save(array: arrayOfStudents)
-    }
+   // }
 }
