--- UniversityHW/StudentAPI.swift
@@ -1,17 +1,0 @@
-//
-//  SWPeople.swift
-//  UniversityHW
-//
-//  Created by Elena Alekseeva on 4/12/20.
-//  Copyright © 2020 Elena Alekseeva. All rights reserved.
-//
-
-import Foundation
-
-// [[String:String]]
-// [[String : Any]]
-
-struct StudentAPI: Codable {
-    let name: String
-}
-
